+++ tests/bfabric_app_runner/inputs/resolve/test_resolve_bfabric_annotation_specs.py
@@ -16,7 +16,7 @@ from bfabric_app_runner.specs.inputs.bfabric_annotation_spec import BfabricAnnot
 
 @pytest.fixture
 def bfabric_instance():
-    return "https://bfabric.example.org/bfabric/"
+    return "https://bfabric.example.org/bfabric"
 
 
 @pytest.fixture