+++ tests/bfabric_cli/test_cli_api_update.py
@@ -10,7 +10,7 @@ from bfabric_scripts.cli.api.output_format import OutputFormat
 @pytest.fixture
 def mock_client(mocker):
     client = mocker.Mock(spec=Bfabric)
-    client.config.base_url = "http://test-bfabric.com"
+    client.config.base_url = "http://test-bfabric.com/bfabric"
     return client